requirements.txt
aioredis==1.0.0
appdirs==1.4.3
asgi-redis==1.4.3
asgiref==2.2.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
channels==2.0.2
channels-redis==2.1.0
constantly==15.1.0
daphne==2.1.0
Django==2.0.3
django-resized==0.3.9
djangorestframework==3.7.7
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
msgpack-python==0.5.5
mysqlclient==1.3.12
Pillow==5.0.0
PyMySQL==0.8.0
pytz==2018.3
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
zope.interface==4.4.3
